Removed shallowEqualScalar, simplified connect update logic
