+++ src/utils/shallowEqual.js
@@ -3,6 +3,11 @@ export default function shallowEqual(objA, objB) {
     return true;
   }
 
+  if (typeof objA !== 'object' || objA === null ||
+      typeof objB !== 'object' || objB === null) {
+    return false;
+  }
+
   const keysA = Object.keys(objA);
   const keysB = Object.keys(objB);
 
@@ -11,10 +16,8 @@ export default function shallowEqual(objA, objB) {
   }
 
   // Test for A's keys different from B.
-  const hasOwn = Object.prototype.hasOwnProperty;
   for (let i = 0; i < keysA.length; i++) {
-    if (!hasOwn.call(objB, keysA[i]) ||
-        objA[keysA[i]] !== objB[keysA[i]]) {
+    if (objA[keysA[i]] !== objB[keysA[i]]) {
       return false;
     }
   }
